updated plotting function
I tried to put all the samples in a big dataframe and cut the reads off in batch. but, different samples have different length distributions. such a method could only brought a big mess. 
in this version, count matrices for the same sample but different cut off values are stored in a list, such a list is stored in a dictionary with the sample name being the key. 
to get the plots done, just pass a list of the names of the samples into the plot function