compare umap with different labels
Added plots for comparing umap processed data with different labels. There are three groups of labels. The first group of labels are the sample names. The second group of labels are the sources of the data, Real or Simulated. The last group of labels are the treatments, High, Low, Veh, Tumor